+++ src/constants/labels.ts
@@ -55,10 +55,12 @@ export const defaultBreadcrumbs = [
 
 export const documentTitles = {
   home: "Digital Research Books Beta | NYPL",
+  advancedSearch: "Advanced Search | Digital Research Books Beta | NYPL",
   search: "Search Results | Digital Research Books Beta | NYPL",
   workItem: "Item Details | Digital Research Books Beta | NYPL",
   editionItem: "Edition Details | Digital Research Books Beta | NYPL",
   readItem: "Read Online | Digital Research Books Beta | NYPL",
+  collection: "Collection | Digital Research Books Beta | NYPL",
 };
 
 export const yearsType = { start: "Start", end: "End" };