+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 - Update to Web Reader version 4.3.1
 - Add TOC for single resource PDFs
 - Refactored Playwright/Cucumber tests to use new dropdown action function
-- Beginning to refactor Playwright/Cucumber tests to use new text fill function
+- Refactored Playwright/Cucumber tests to use new text fill function
 
 ## [0.17.1]
 - Replace dgx-header-component and DS Footer with Header/Footer components from nypl-header-app