+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - Refactored Playwright/Cucumber tests to use new dropdown action function
 - Refactored Playwright/Cucumber tests to use new text fill function
 - Migrate from Google Analytics to Adobe Analytics
-- Preparing to refactor Playwright/Cucumber tests to use new display function
+- Refactored Playwright/Cucumber tests to use new visibility function
 
 ## [0.17.1]
 - Replace dgx-header-component and DS Footer with Header/Footer components from nypl-header-app